+++ sample-tom/src/main/java/com/usthe/sureness/sample/tom/service/impl/RoleServiceImpl.java
@@ -4,9 +4,11 @@ import com.usthe.sureness.sample.tom.dao.AuthRoleDao;
 import com.usthe.sureness.sample.tom.pojo.entity.AuthRoleDO;
 import com.usthe.sureness.sample.tom.service.RoleService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Example;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
@@ -16,6 +18,7 @@ import java.util.Optional;
  * @date 13:10 2019-08-04
  */
 @Service
+@Transactional(rollbackFor = Exception.class)
 public class RoleServiceImpl implements RoleService {
 
     @Autowired
@@ -23,23 +26,39 @@ public class RoleServiceImpl implements RoleService {
 
     @Override
     public boolean isRoleExist(AuthRoleDO authRole) {
-        return authRoleDao.existsById(authRole.getId());
+        AuthRoleDO role = AuthRoleDO.builder()
+                .name(authRole.getName()).code(authRole.getCode()).build();
+        return authRoleDao.exists(Example.of(role));
     }
 
     @Override
     public boolean addRole(AuthRoleDO authRole) {
-        return authRoleDao.save(authRole) != null;
+        if (isRoleExist(authRole)) {
+            return false;
+        } else {
+            authRoleDao.saveAndFlush(authRole);
+            return true;
+        }
     }
 
     @Override
     public boolean updateRole(AuthRoleDO authRole) {
-        return authRoleDao.save(authRole) != null;
+        if (authRoleDao.existsById(authRole.getId())) {
+            authRoleDao.saveAndFlush(authRole);
+            return true;
+        } else {
+            return false;
+        }
     }
 
     @Override
     public boolean deleteRole(Long roleId) {
-        authRoleDao.deleteById(roleId);
-        return true;
+        if (authRoleDao.existsById(roleId)) {
+            authRoleDao.deleteById(roleId);
+            return true;
+        } else {
+            return false;
+        }
     }
 
     @Override
@@ -49,9 +68,8 @@ public class RoleServiceImpl implements RoleService {
     }
 
     @Override
-    public Optional<Page<AuthRoleDO>> getPageRole(Integer currentPage, Integer pageSize) {
+    public Page<AuthRoleDO> getPageRole(Integer currentPage, Integer pageSize) {
         PageRequest pageRequest = PageRequest.of(currentPage, pageSize);
-        Page<AuthRoleDO> page = authRoleDao.findAll(pageRequest);
-        return Optional.of(page);
+        return authRoleDao.findAll(pageRequest);
     }
 }
